+++ app/ante_no_seq.go
@@ -8,7 +8,9 @@ import (
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	ibcante "github.com/cosmos/ibc-go/v7/modules/core/ante"
 	ibckeeper "github.com/cosmos/ibc-go/v7/modules/core/keeper"
+	"github.com/skip-mev/block-sdk/block"
 
+	"github.com/cosmos/cosmos-sdk/client"
 	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	ante "github.com/cosmos/cosmos-sdk/x/auth/ante"
@@ -21,12 +23,26 @@ import (
 
 	txfeeskeeper "github.com/osmosis-labs/osmosis/v24/x/txfees/keeper"
 	txfeestypes "github.com/osmosis-labs/osmosis/v24/x/txfees/types"
+	auctionkeeper "github.com/skip-mev/block-sdk/x/auction/keeper"
+
+	auctionante "github.com/skip-mev/block-sdk/x/auction/ante"
+
+	txfeeskeeper "github.com/osmosis-labs/osmosis/v24/x/txfees/keeper"
+	txfeestypes "github.com/osmosis-labs/osmosis/v24/x/txfees/types"
 )
+
+// BlockSDKAnteHandlerParams are the parameters necessary to configure the block-sdk antehandlers
+type BlockSDKAnteHandlerParams struct {
+	freeLane      block.Lane
+	mevLane       auctionante.MEVLane
+	auctionKeeper auctionkeeper.Keeper
+	txConfig      client.TxConfig
+}
 
 // Link to default ante handler used by cosmos sdk:
 // https://github.com/cosmos/cosmos-sdk/blob/v0.43.0/x/auth/ante/ante.go#L41
 // N.B. There is a sister file called `ante_eq.go` that is used for production.
-// Opposite to this, it also chains `IncrementSequenceDecorator` which is not needed for e2e testing but is critical for production.
+// Opposite to this, it also chains `SigVerificationDecorator` which is not needed for e2e testing but is critical for production.
 // If you make a change here, make sure to make the same change in `ante.go`.
 func NewAnteHandler(
 	appOpts servertypes.AppOptions,
@@ -39,6 +55,7 @@ func NewAnteHandler(
 	sigGasConsumer ante.SignatureVerificationGasConsumer,
 	signModeHandler signing.SignModeHandler,
 	channelKeeper *ibckeeper.Keeper,
+	blockSDKParams BlockSDKAnteHandlerParams,
 ) sdk.AnteHandler {
 	mempoolFeeOptions := txfeestypes.NewMempoolFeeOptions(appOpts)
 	mempoolFeeDecorator := txfeeskeeper.NewMempoolFeeDecorator(*txFeesKeeper, mempoolFeeOptions)
@@ -59,12 +76,21 @@ func NewAnteHandler(
 		ante.TxTimeoutHeightDecorator{},
 		ante.NewValidateMemoDecorator(ak),
 		ante.NewConsumeGasForTxSizeDecorator(ak),
-		deductFeeDecorator,
+		block.NewIgnoreDecorator(
+			deductFeeDecorator,
+			blockSDKParams.freeLane,
+		),
 		ante.NewSetPubKeyDecorator(ak), // SetPubKeyDecorator must be called before all signature verification decorators
 		ante.NewValidateSigCountDecorator(ak),
 		ante.NewSigGasConsumeDecorator(ak, sigGasConsumer),
-		ante.NewSigVerificationDecorator(ak, signModeHandler),
-		// ante.NewIncrementSequenceDecorator(ak),
+		// ante.NewSigVerificationDecorator(ak, signModeHandler) <-- removed this to prevent failures resulting from invalid tx orders in e2e
+		ante.NewIncrementSequenceDecorator(ak),
 		ibcante.NewRedundantRelayDecorator(channelKeeper),
+		// auction module antehandler
+		auctionante.NewAuctionDecorator(
+			blockSDKParams.auctionKeeper,
+			blockSDKParams.txConfig.TxEncoder(),
+			blockSDKParams.mevLane,
+		),
 	)
 }