+++ app/keepers/modules.go
@@ -61,6 +61,7 @@ import (
 	valsetprefmodule "github.com/osmosis-labs/osmosis/v24/x/valset-pref/valpref-module"
 	"github.com/osmosis-labs/osmosis/x/epochs"
 	ibc_hooks "github.com/osmosis-labs/osmosis/x/ibc-hooks"
+	"github.com/skip-mev/block-sdk/x/auction"
 )
 
 // AppModuleBasics returns ModuleBasics for the module BasicManager.
@@ -130,4 +131,5 @@ var AppModuleBasics = []module.AppModuleBasic{
 	packetforward.AppModuleBasic{},
 	cosmwasmpoolmodule.AppModuleBasic{},
 	tendermint.AppModuleBasic{},
+	auction.AppModuleBasic{},
 }